Corrections in ACL/ACE Abuses Descriptions (#262)
* Update General.tsx

Changed `privilege` to `permission` for the description of `GenericAll`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update LinuxAbuse.tsx

Changed `privilege` to `permission` for the description of `GenericAll`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update Opsec.tsx

Changed `privilege` to `permission` for the description of `GenericAll`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update WindowsAbuse.tsx

Changed `privilege` to `permission` for the description of `GenericAll`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update LinuxAbuse.tsx

Changed `privilege` to `permission` for `GenericWrite`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update Opsec.tsx

Changed `privilege` to `permission` for `GenericWrite`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update WindowsAbuse.tsx

Changed `privilege` to `permission` for `GenericWrite`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update LinuxAbuse.tsx

Changed `privilege` to `permission` for `WriteOwner`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update Opsec.tsx

Changed `privilege` to `permission` for `WriteOwner`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update WindowsAbuse.tsx

Changed `privilege` to `permission` for `WriteOwner`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update General.tsx

Changed `privilege` to `permission` for `WriteDACL`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update LinuxAbuse.tsx

Changed `privilege` to `permission` for `WriteDACL`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update Opsec.tsx

Changed `privilege` to `permission` for `WriteDACL`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update WindowsAbuse.tsx

Changed `privilege` to `permission` for `WriteDACL`. The attack is an Abuse of Active Directory ACLs/ACEs consisting of permissions.

* Update General.tsx

Changed the term "privilege" to "permission" for the DCSync abuse description, since DCSync is an account permission and not a privilege.

---------

Co-authored-by: John Hopper <jhopper@specterops.io>